Refactoring
Refactoring:
- Outsourcing the XMLfile reading.
This adds complexity to ExercisePage to ask it to read a file. The exercise page job should just do the exercise with the user, not have to read a file. So from now on, it will receive the data structure linked to the exercise and the XML exercise reading will be done upstream.
-
Other changes in ExercisePage to improve readability