feat(CheckPosition):  improved checkposition clarity and removing min/max angle constraint
CheckPosition: improve clarity by drawing angle arc with multiple lines.
The arc of a circle must be composed of several lines to have a round shape.

Additionally, the constraint requiring the minimum angle to be less than the maximum angle has been removed. This change is necessary to accommodate angle ranges that cross the 0° point, such as angles from 150° to 140°. Removing this constraint ensures that the function can handle these cases correctly and display the angle range as intended.